third_party/blink/web_tests/external/wpt/css/css-transitions/changing-while-transition-004-expected.txt

This is a testharness.js-based test.
[FAIL] Unsetting transition should not affect in-flight transitions
  assert_equals: Even after unsetting transition, the transition should be 50% complete expected "50px" but got "100px"
Harness: the test ran to completion.
